requirements.txt
certifi==2018.1.18
chardet==3.0.4
chardet2==2.0.3
click==6.7
cryptography==1.7.1
Flask==0.12.2
Flask-WTF==0.14.2
future==0.16.0
get==0.0.39
idna==2.6
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.0
post==0.0.26
public==0.0.65
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pyserial==3.4
python-apt==1.1.0b5
pyxdg==0.25
PyYAML==3.12
query-string==0.0.28
request==0.0.26
requests==2.12.4
RPi.GPIO==0.6.3
SecretStorage==2.3.1
serial==0.0.38
six==1.10.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
